Move model classes into model package
