+++ src/data/model/DataAddress.java
@@ -1,4 +1,4 @@
-package data;
+package data.model;
 
 import acq.IAddress;
