+++ src/data/model/DataCase.java
@@ -1,4 +1,4 @@
-package data;
+package data.model;
 
 import acq.ICase;
 import acq.ICitizen;
