+++ src/data/model/DataCitizen.java
@@ -1,9 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package data;
+package data.model;
 
 import acq.IAddress;
 import acq.ICitizen;
